refactor(mobile): 역 상세 이동 안내 컴포넌트 분리
